accelerants: Fix warning introduced in hrev48265
* Align types for compare
* Fixes builds

+++ src/add-ons/accelerants/s3/savage_mode.cpp
@@ -565,7 +565,7 @@ Savage_WriteMode(const DisplayModeEx& mode, const SavageRegRec& regRec)
 
 	WriteCrtcReg(0x11, 0x00, 0x80);		// unlock CRTC reg's 0-7 by clearing bit 7 of cr11
 
-	for (int j = 0; j < B_COUNT_OF(regRec.CRTC); j++)
+	for (int j = 0; j < (int)B_COUNT_OF(regRec.CRTC); j++)
 		WriteCrtcReg(j, regRec.CRTC[j]);
 
 	// Setup HSYNC & VSYNC polarity.
